Refactor File management environments initialization
