+++ GliaWidgets/Sources/Storage/FileSystemStorage.Environment.Interface.swift
@@ -7,3 +7,21 @@ extension FileSystemStorage {
         var date: () -> Date
     }
 }
+
+extension FileSystemStorage.Environment {
+    static func create(with environment: FileDownloader.Environment) -> Self {
+        .init(
+            fileManager: environment.fileManager,
+            data: environment.data,
+            date: environment.date
+        )
+    }
+
+    static func create(with environment: FileUploader.Environment) -> Self {
+        .init(
+            fileManager: environment.fileManager,
+            data: environment.data,
+            date: environment.date
+        )
+    }
+}
