+++ GliaWidgets/Sources/Upload/FileUploader.Environment.Interface.swift
@@ -16,3 +16,31 @@ extension FileUploader {
 extension FileUploader.Environment {
     typealias UploadFile = FileUpload.Environment.UploadFile
 }
+
+extension FileUploader.Environment {
+    static func create(with environment: SecureConversations.Coordinator.Environment) -> Self {
+        .init(
+            uploadFile: .toSecureMessaging(environment.uploadSecureFile),
+            fileManager: environment.fileManager,
+            data: environment.data,
+            date: environment.date,
+            gcd: environment.gcd,
+            uiScreen: environment.uiScreen,
+            createThumbnailGenerator: environment.createThumbnailGenerator,
+            uuid: environment.uuid
+        )
+    }
+
+    static func create(with environment: SecureConversations.TranscriptModel.Environment) -> Self {
+        .init(
+            uploadFile: .toSecureMessaging(environment.secureUploadFile),
+            fileManager: environment.fileManager,
+            data: environment.data,
+            date: environment.date,
+            gcd: environment.gcd,
+            uiScreen: environment.uiScreen,
+            createThumbnailGenerator: environment.createThumbnailGenerator,
+            uuid: environment.uuid
+        )
+    }
+}
